Eliminate warning from eslint
