+++ .eslintrc.yml
@@ -1,0 +1,32 @@
+extends: 'airbnb'
+
+env:
+  browser: true
+
+rules:
+  semi: off
+  space-before-function-paren: [ error, never ]
+  comma-dangle: off
+  no-multi-spaces: [ error, exceptions: { VariableDeclarator: true } ]
+  newline-per-chained-call: off
+  prefer-arrow-callback: off
+  max-len: [ warn, 150 ]
+  object-shorthand: warn
+  func-names: off
+  import/no-unresolved: off
+  no-param-reassign: [warn, { props: false }]
+  react/no-danger: off
+  class-methods-use-this: off
+  import/extensions: [ warn, { jsx: "never" } ]
+  jsx-a11y/no-static-element-interactions: off
+  jsx-a11y/media-has-caption: off
+  import/no-webpack-loader-syntax: off
+  import/no-extraneous-dependencies: warn
+  react/require-default-props: off
+  react/no-render-return-value: off
+
+  react/forbid-prop-types: off
+  react/prop-types: warn
+  react/no-unused-prop-types: warn
+  react/jsx-curly-brace-presence: off
+
