+++ app/javascript/bootstrap-v4-ext/index.jsx
@@ -1,12 +1,14 @@
 import $ from 'jquery'
 import './all.scss'
 
-$(document).on('content:loaded', function(event) {
-  $(event.target).on('click', '.remove-attachment', function(event) {
-    const $currentTarget = $(event.currentTarget)
-    const $attachment = $($currentTarget.data('target'))
+function removeAttachment(event) {
+  const $currentTarget = $(event.currentTarget)
+  const $attachment = $($currentTarget.data('target'))
 
-    $attachment.hide()
-    $attachment.find('.destroy-task-attachment').val("1")
-  })
+  $attachment.hide()
+  $attachment.find('.destroy-task-attachment').val('1')
+}
+
+$(document).on('content:loaded', function(event) {
+  $(event.target).on('click', '.remove-attachment', removeAttachment)
 })
