+++ app/javascript/components/drag_and_drop.jsx
@@ -1,28 +1,5 @@
 import $ from 'jquery'
 
-$(function() {
-  $(document).find('#uploadAttachmentsModal').on('dragstart drag dragenter dragexit dragleave dragover dragend drop', function(event) {
-    event.preventDefault()
-    event.stopPropagation()
-  }).on('dragenter dragover', function(event) {
-    const $targetArea = $(event.currentTarget)
-    $targetArea.find('#uploadAttachmentArea').removeClass('border-white bg-light').addClass('border-secondary')
-  }).on('dragleave', function(event) {
-    const $targetArea = $(event.currentTarget)
-    $targetArea.find('#uploadAttachmentArea').removeClass('border-secondary').addClass('border-white bg-light')
-  }).on('drop', function(event) {
-    const $targetArea = $(event.currentTarget)
-    $targetArea.find('#uploadAttachmentArea').removeClass('border-secondary').addClass('border-white bg-light')
-
-    const requestParams = $targetArea.data()
-    const files = event.originalEvent.dataTransfer.files
-
-    if (files.length >= 1) {
-      uploadAttachment($targetArea.find('#uploadAttachmentArea'), files[0], requestParams)
-    }
-  })
-})
-
 function uploadAttachment(target, file, { url, method }) {
   const formData = new FormData()
   formData.append('attachment[file]', file)
@@ -41,3 +18,26 @@ function uploadAttachment(target, file, { url, method }) {
     target.before(respond)
   })
 }
+
+$(function() {
+  $(document).find('#uploadAttachmentsModal').on('dragstart drag dragenter dragexit dragleave dragover dragend drop', function(event) {
+    event.preventDefault()
+    event.stopPropagation()
+  }).on('dragenter dragover', function(event) {
+    const $targetArea = $(event.currentTarget)
+    $targetArea.find('#uploadAttachmentArea').removeClass('border-white bg-light').addClass('border-secondary')
+  }).on('dragleave', function(event) {
+    const $targetArea = $(event.currentTarget)
+    $targetArea.find('#uploadAttachmentArea').removeClass('border-secondary').addClass('border-white bg-light')
+  }).on('drop', function(event) {
+    const $targetArea = $(event.currentTarget)
+    $targetArea.find('#uploadAttachmentArea').removeClass('border-secondary').addClass('border-white bg-light')
+
+    const requestParams = $targetArea.data()
+    const { files } = event.originalEvent.dataTransfer
+
+    if (files.length >= 1) {
+      uploadAttachment($targetArea.find('#uploadAttachmentArea'), files[0], requestParams)
+    }
+  })
+})
