+++ app/javascript/components/flatpickr/selectToday.scss
