+++ app/javascript/components/vendor.jsx
@@ -1,0 +1,4 @@
+import $ from 'jquery'
+import 'jquery-ujs'
+
+window.$ = $
