--- app/javascript/widgets/flatpickr/index.jsx
@@ -1,22 +1,0 @@
-import $ from 'jquery'
-
-import 'flatpickr'
-import 'flatpickr/dist/flatpickr.css'
-import 'flatpickr/dist/l10n'
-
-import confirmDatePlugin from 'flatpickr/dist/plugins/confirmDate/confirmDate'
-import 'flatpickr/dist/plugins/confirmDate/confirmDate.css'
-
-import selectTodayPlugin from  './selectToday'
-import './selectToday.scss'
-
-$(document).on('turbolinks:load', function() {
-    $(".datetime-picker").each(function() {
-        $(this).flatpickr({
-            enableTime: true,
-            wrap: true,
-            allowInput: true,
-            plugins: [ new selectTodayPlugin(), new confirmDatePlugin() ]
-        })
-    })
-})
