+++ app/views/issues/show.html.erb
@@ -35,7 +35,6 @@
         <% end %>
       </div>
     </div>
-
   </div>
   <hr>
 <% end %>
@@ -56,10 +55,3 @@
 <% end %>
 
 <%= render "comments/form", issue: @issue %>
-
-<div id="add_label" class="modal">
-  <div class="modal-dialog modal-lg">
-    <div class="modal-content">
-    </div>
-  </div>
-</div>
