+++ package.json
@@ -15,6 +15,11 @@
     "turbolinks": "^5.0.3"
   },
   "devDependencies": {
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.9.0",
     "webpack-dev-server": "2.11.2"
   }
 }
